+++ src/main/java/org/vaadin/example/application/views/DetailedDepotView.java
@@ -293,6 +293,5 @@ public class DetailedDepotView extends AbstractSideNav implements HasUrlParamete
         dialog.open();
     }
 }
-//TODO:Einbinden der Funktionalität zum Kaufen und Verkaufen von Wertpapieren
-//TODO:Wertpapiere in die Depot-Übersicht einfügen
+
